Silence "Killed" messages from shtk_process_run -t
Some shells (NetBSD's sh or bash) output notifications for processes
that do not exit cleanly when using wait on them.  The new timeout
feature for subprocess execution always generates such notifications
because either the process being run is killed or the background
watchdog process is killed.

These notifications are annoying because they are expected to happen, so
just silence them.  As a bonus point, we fix two tests.