Fix distcheck due to version.subr breakage
Because version.subr is a generated file and thus it exists only
in the build directory, distcheck broke.  Change the build to pass
the correct path to the build directory so that version.subr can
be found.  Just abuse SHTK_MODULESPATH for this.

While doing this, also ensure that tests for shtk_import do not
inadvertently use the value of SHTK_MODULESPATH value.  Wasn't a
problem before because this variable was undefined but it became
one now.