Treat returning from setup/test/teardown as a pass
If setup/test/teardown evaluated a command as false as their very last
statement, the test case would be misteriously marked as failed.  This
is unexpected and surprising (especially because we do not "set -e")
and thus should not happen.

Change the code to discard the value of the last evaluated statement
so that, if the execution falls through the end of setup, test and
teardown methods, the test case is considered a success.

These routines can still "exit 1" (through "fail") to mark a failure.