Support debug messages and choosing a log level
This change adds a new shtk_cli_debug function to support generating
debug-level messages: i.e. messages that are useful to debug a problem
but that should not be shown during regular usage.

To complement the previous, add functionality to allow choosing the
desired log level by means of the shtk_cli_set_log_level function and
to query what the current level is with the shtk_cli_log_level function.