Fix pattern matching in shtk_list_filter
We cannot just pass a variable's contents as the pattern in a case branch
as this does not work in all cases: in particular, alternative branches do
not seem to work in any shell and pdksh does not recognize any pattern.

Fix this by running an eval on the case statement.