Ignore signals in cleanup tests for pdksh
pdksh seems to differ in how it handles interactive termination signals:
when our tests send SIGINT to themselves, the signal ends bubbling up to
the parent shells (even if the parent shell is not pdksh!).  This seems
to imply that pdksh is resending the signal to the process group.

I can't tell if this is correct behavior or if it is a bug in pdksh.
http://www.cons.org/cracauer/sigint.html does not pain a pretty
picture on how SIGINT handling works and, given its complexity, I'd
expect different implementations to do different things.

Therefore, just appease pdksh by ignoring the signals during our tests.
Programs using shtk_cleanup will hopefully do the right thing as well.