Mark cleanup_test:register__sighup as xfail on zsh
zsh seems to do something funny with the propagation of SIGHUP to the
caller: it always exits with 1 instead of reporting the signal.

The documentation is silent on this topic so I cannot tell if this is
expected or if this is a bug in zsh.  Therefore, just mark this one
particular test as an expected failure on zsh.