Fix shtk_abort on NetBSD 7.0 as of today
It turns out that "ps $$" on NetBSD properly filters the output to the
given PID, but "ps -o ppid $$" does not.  Make sure to always pass the
PID via the -p flag, which is what the manpage says one must do.

I'm pretty sure this used to work on NetBSD, so I suspect ps has been
broken by any of the recent changes to it.  Anyway, better use the
documented syntax than the lone PID argument.