Add a website for shtk
This creates a minimal website for shtk with the main goal of hosting
the reference manual.  Otherwise it's impossible for potential users
to see all of the features included in this library.

This is intentionally simple.  I'm choosing to just create my own
bespoke templating system to generate the pages given that most of
the work here is integrating with mandoc.