Add basic implementation of rules_shtk for Bazel
This change brings a new implementation of shtk build rules for Bazel,
namely rules_shtk.

It does so by: adding a new bazel_rules subdirectory, which contains the
code of the rules_shtk; adding a new bazel_examples subdirectory, which
provides end-to-end examples for usage of rules_shtk; and adding a new
integration test (inttest) that leverages the bazel_examples to verify
that the rules work as expected.